==> all-features/1-partition.ipynb <==
# Assign positives and negatives 

import itertools
import random
import sys
import json

import py2neo
import pandas

## Startup neo4j and connections

with open('servers.json') as read_file:
    instances = json.load(read_file)

for instance in instances:
    uri = 'http://localhost:{}/db/data/'.format(instance['port'])
    instance['py2neo'] = py2neo.Graph(uri)
    if instance['name'] == 'rephetio-v2.0':
        neo_unperm = instance['py2neo']

hetnets = [x['name'] for x in instances]
hetnets

## Create partitions

def to_df(record_list):
    """Convert a py2neo RecordList to a dataframe"""
    return pandas.DataFrame(record_list.records, columns = record_list.columns)

compound_df = pandas.read_table('../summary/compounds.tsv')
disease_df = pandas.read_table('../summary/diseases.tsv')

'{} compounds × {} diseases = {} pairs'.format(
    len(disease_df), len(compound_df), len(disease_df) * len(compound_df))

nonzero_prior_pairs = set(itertools.product(
    compound_df.query("treats > 0").compound_id,
    disease_df.query("treats > 0").disease_id)
)

indication_query = '''
MATCH (compound:Compound)-[treatment:TREATS_CtD]->(disease:Disease)
RETURN
  compound.identifier AS compound_id,
  disease.identifier AS disease_id
ORDER BY
  compound_id, disease_id
'''

def partition(neo):
    """
    Extract negative and positive compound-disease pairs from a py2neo.Graph.
    """
    # Use TREATS_CtD as positives
    indication_df = to_df(neo.cypher.execute(indication_query))
    positives = set(zip(indication_df.compound_id, indication_df.disease_id))
    # Use nonzero-prior pairs excluding non-negatives as negatives
    negatives = nonzero_prior_pairs - positives
    negatives = random.sample(negatives, k=len(positives) * 4)
    rows = list()
    for status, pairs in (0, negatives), (1, positives):
        for drug, disease in pairs:
            rows.append((drug, disease, status))
    df = pandas.DataFrame(rows, columns=['compound_id', 'disease_id', 'status'])
    df = df.sort_values(['disease_id', 'status', 'compound_id'])
    return df

part_dfs = list()
for seed, instance in enumerate(instances):
    random.seed(seed, version=2)
    part_df = partition(instance['py2neo'])
    part_df.insert(0, 'hetnet', instance['name'])
    part_dfs.append(part_df)
part_df = pandas.concat(part_dfs)

part_df.head(2)

## Create an even-matrix for unpermuted observations
template_df = pandas.DataFrame(list(itertools.product(hetnets, compound_df.compound_id)), columns=['hetnet', 'compound_id'])
unperm_pair_df = part_df.query("hetnet == 'rephetio-v2.0'")[['compound_id', 'disease_id']]
unperm_pair_df = template_df.merge(unperm_pair_df)
unperm_pair_df = unperm_pair_df.merge(part_df, how='left')
unperm_pair_df = unperm_pair_df[unperm_pair_df.status.isnull()]
unperm_pair_df.status = unperm_pair_df.status.fillna(0).astype(int)
unperm_pair_df.head(2)

part_df['primary'] = 1
unperm_pair_df['primary'] = 0
full_part_df = pandas.concat([part_df, unperm_pair_df])
full_part_df = full_part_df.sort_values(['compound_id', 'disease_id', 'hetnet'])
full_part_df.head(5)

full_part_df.to_csv('data/partitions.tsv', sep='\t', index=False)

# Number of hetnet-compound-disease pairs
len(full_part_df)

# Number of positives and negatives per purpose
pandas.crosstab(full_part_df.primary, full_part_df.status)

# Number of positives and negatives per hetnet
pandas.crosstab(full_part_df.hetnet, full_part_df.status)
