==> all-features/8-model-performances.ipynb <==
# Compare model performance between real and permuted hetnets

library(magrittr)

auroc_df = readr::read_tsv('data/auroc.tsv')
degrees = dplyr::filter(auroc_df, feature_type == 'degree')$feature
metapaths = dplyr::filter(auroc_df, feature_type == 'dwpc')$feature

# col_types not needed here, but used for safety
col_types = list()
for (metapath in metapaths) {
  col_types[[metapath]] = readr::col_number()
}
for (degree in degrees) {
  col_types[[degree]] = readr::col_integer()
}

feature_df = readr::read_tsv('data/matrix/features.tsv.bz2', col_types = col_types)

transform_df = function(df) {
    
  df = dplyr::bind_cols(
    df %>%
    dplyr::transmute(status, prior_logit = boot::logit(prior_prob)),

    df %>%
    dplyr::select(one_of(degrees)) %>%
    dplyr::mutate_each(dplyr::funs(asinh)),

    df %>%
    dplyr::select(one_of(metapaths)) %>%
    dplyr::mutate_each(dplyr::funs(asinh(. / mean(.))))
  )
  return(df)
}

transformed_df = feature_df %>%
  dplyr::group_by(hetnet) %>%
  dplyr::do(transform_df(.)) %>%
  dplyr::ungroup()

head(transformed_df, 2)

fit_list = list()
i = 0
get_performance = function(df, incl_degrees) {
  for (seed in 1:5) {
    for (alpha in 0:1) {
      i <<- i + 1
      y = df$status
      X = df %>% dplyr::select(-status, -hetnet) %>% as.matrix()
      penalty_factor = ifelse(colnames(X) == 'prior_logit', 0, 1)
      fit = hetior::glmnet_train(X = X, y = y, alpha = alpha, cores = 5, seed=seed,
        penalty.factor=penalty_factor, lambda.min.ratio=1e-6, nlambda=200
      )
      fit$name = df$hetnet[1]
      fit$incl_degrees = incl_degrees
      fit_list[[i]] <<- fit
    }
  }
  return(data.frame(i))
}

temp = transformed_df %>%
  dplyr::group_by(hetnet) %>%
  dplyr::do(get_performance(., incl_degrees=1))

temp = transformed_df %>%
  dplyr::select(-one_of(degrees)) %>%
  dplyr::group_by(hetnet) %>%
  dplyr::do(get_performance(., incl_degrees=0))

result_df = fit_list %>% lapply(function(l) {
  dplyr::data_frame(
    name = l$name,
    alpha = l$alpha,
    incl_degrees = l$incl_degrees,
    seed = l$seed,
    auroc = l$vtm$auroc,
    auprc = l$vtm$auprc,
    tjur = l$vtm$tjur
  )
    
}) %>% dplyr::rbind_all()

head(result_df, 2)

result_df %>% readr::write_tsv('data/model-performances.tsv')

summary_df = result_df %>%
  dplyr::mutate(permuted = as.integer(grepl('perm', result_df$name))) %>%
  dplyr::group_by(permuted, alpha, incl_degrees, alpha) %>%
  dplyr::do(
    dplyr::bind_cols(
      ggplot2::mean_cl_normal(.$tjur) %>% dplyr::rename(tjur=y, tjur_lower=ymin, tjur_upper=ymax),
      ggplot2::mean_cl_normal(.$auroc) %>% dplyr::rename(auroc=y, auroc_lower=ymin, auroc_upper=ymax),
      ggplot2::mean_cl_normal(.$auprc) %>% dplyr::rename(auprc=y, auprc_lower=ymin, auprc_upper=ymax))
  ) %>%
  dplyr::ungroup() %>%
  dplyr::arrange(desc(tjur))

summary_df %>% readr::write_tsv('data/model-performances-summary.tsv')
summary_df
